Refactor contentFinder into searchableEditorPopup
